% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{analyse_mr}
\alias{analyse_mr}
\title{Analyse Mark-Recapture Data}
\usage{
analyse_mr(x, niters = 10^5)
}
\arguments{
\item{x}{A data.frame of the mark-recapture data to analyse.}

\item{niters}{An integer of the minimum number of MCMC iterations to
perform.}
}
\value{
A jags_analysis object.
}
\description{
Analyses mark-recapture data using a Bayesian Jolly-Seber
model to estimate the abundance, annual survival and (re)capture probability.
The data must be in the same format as the \code{\link{ferox}}
data provided with this package (only the Fish and Date columns are required).
}
\seealso{
\code{\link{ranmr}}
}

