% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ferox}
\alias{ferox}
\title{Ferox Trout Mark-Recapture Data}
\usage{
ferox()
}
\value{
A data.frame of ferox trout mark-recapture data from
the \code{ranmrdata} package.
}
\description{
Ferox Trout Mark-Recapture Data
}
\examples{
ferox()
}
\seealso{
\code{ranmrdata::\link[ranmrdata]{ferox}}
}

