% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fill_in_ages}
\alias{fill_in_ages}
\title{Fill In Ages}
\usage{
fill_in_ages(x)
}
\arguments{
\item{x}{The data.frame with missing ages.}
}
\value{
The modified data frame.
}
\description{
Fill In Ages
}
\examples{
ferox <- ranmrdata::ferox
fill_in_ages(ferox@data)
}

