% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{mr_model_code}
\alias{mr_model_code}
\title{Mark-Recapture Model Code}
\usage{
mr_model_code()
}
\value{
A string of the JAGS model code.
}
\description{
Returns a string of the JAGS code
defining the mark-recapture model.
}
\examples{
cat(mr_model_code())
}
\seealso{
\code{\link{ranmr}}
}

