% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot Mark-Recapture Analysis}
\usage{
plot_abundance(x)
}
\arguments{
\item{x}{A jags_analysis object of the mark-recapture analysis to plot.}
}
\value{
A ggplot object.
}
\description{
Plots a mark-recapture analysis as an abundance plot.
}
\seealso{
\code{\link{ranmr}}
}

