% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mr}
\alias{plot_mr}
\title{Plot Mark-Recapture Data}
\usage{
plot_mr(x, xcol = "Length", ycol = "Mass", xlab = NULL, ylab = NULL,
  gp = NULL)
}
\arguments{
\item{x}{A data.frame of the mark-recapture data to plot.}

\item{xcol}{A string of the column to plot on the x-axis.}

\item{ycol}{A string of the column to plot on the y-axis.}

\item{xlab}{A optional string of the x-axis label.}

\item{ylab}{A optional string of the y-axis label.}

\item{gp}{An optional ggplot object}
}
\value{
A ggplot object.
}
\description{
Plots mark-recapture data as an x-y scatterplot where consecutive
encounters with the same individual are joined by lines.
The data must be in the same format as the \code{\link{ferox}}
data provided with this package (only the two columns specified in the
plot plus the Fish and Date columns are required).
}
\examples{
plot_mr(ferox())
}
\seealso{
\code{\link{ranmr}}
}

