% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rannoch}
\alias{plot_rannoch}
\title{Plot Loch Rannoch}
\usage{
plot_rannoch(rannoch)
}
\arguments{
\item{rannoch}{A data frame of the rannoch polygons with columns Easting and Northing in km.}
}
\value{
A ggplot2 object.
}
\description{
Returns a ggplot2 object of a map of Loch Rannoch.
}
\examples{
rannoch <- ggplot2::fortify(ranmrdata::rannoch)
rannoch <- dplyr::mutate(rannoch, Easting = long / 1000, Northing = lat / 1000)
plot_rannoch(rannoch)
}

