% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ranmr}
\alias{ranmr}
\alias{ranmr-package}
\title{Loch Rannoch Ferox Trout Abundance}
\description{
\code{ranmr} is the companion R package for Thorne et al. (xx).
It provides the data and analytic methods and generates the manuscript.
}
\details{
To print a vector of all the exported objects
including functions and data
type \code{objects("package:ranmr")}.
For more information on a particular object
type \code{?object_name}.
The provided data objects are described by \code{data(package = "ranmr")}.
}

