% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml.R
\name{save_kml}
\alias{save_kml}
\title{Save KML File}
\usage{
save_kml(x, file = "ferox.kml")
}
\arguments{
\item{x}{A data.frame of the mark-recapture data to plot.}

\item{file}{A string of the file name.}
}
\description{
Saves mark-recapture data as a KML file for viewing in google earth
where consecutive
encounters with the same individual are joined by lines.
The data must be in the same format as the \code{\link{ferox}}
data provided with this package.
}
\seealso{
\code{\link{ranmr}}
}

