% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tabulate_mr}
\alias{tabulate_mr}
\title{Tabulate Mark-Recapture Data}
\usage{
tabulate_mr(x)
}
\arguments{
\item{x}{A data.frame of the mark-recapture data to tabulate.}
}
\value{
A matrix of the numbers of captures and recaptures by year.
}
\description{
Tabulates mark-recapture data as a matrix.
The data must be in the same format as the \code{\link{ferox}}
data provided with this package (only the Fish and Date columns are required).
}
\examples{
tabulate_mr(ferox())
}
\seealso{
\code{\link{ranmr}}
}

