% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summarise_mr}
\alias{summarise_mr}
\title{Summarise Mark-Recapture Data}
\usage{
summarise_mr(x)
}
\arguments{
\item{x}{A data.frame of the mark-recapture data to summarise.}
}
\value{
A named list of the summary values.
}
\description{
Summarises mark-recapture data as a named list.
The data must be in the same format as the \code{\link{ferox}}
data provided with this package (only the Fish and Date columns are required).
}
\examples{
summarise_mr(ranmrdata::ferox@data)
}
\seealso{
\code{\link{ranmr}}
}

